webscraping partially done, bot not yet fully implemented
Almost all webscraping code has been done, fotocasa and idealista do not fully work yet, but the rest of the code is done. The bot is not yet implemented, but it has already been created.

Only thing left is give the bot a "GUI" and make able to give alerts corresponding to new available houses.